requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
boto3==1.28.0
botocore==1.31.85
celery==5.3.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==4.2
django-redis==5.3.0
djangorestframework==3.14.0
flower==2.0.0
humanize==4.15.0
idna==3.11
iniconfig==2.3.0
jmespath==1.1.0
kombu==5.6.2
packaging==26.0
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
redis==5.0.0
requests==2.31.0
s3transfer==0.6.2
six==1.17.0
sqlparse==0.5.5
tornado==6.5.4
tzdata==2025.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.6.0
